+++ code/lib/core/i18n/l10n/app_localizations_it.dart
@@ -401,6 +401,9 @@ class AppLocalizationsIt extends AppLocalizations {
   String get settingsCodeFontFamily => 'Font del codice';
 
   @override
+  String get settingsEditorFontFamily => 'Font dell\'editor';
+
+  @override
   String get settingsEditorMaxWidth => 'Larghezza massima editor';
 
   @override