+++ code/lib/core/i18n/l10n/app_localizations_ja.dart
@@ -400,6 +400,9 @@ class AppLocalizationsJa extends AppLocalizations {
   String get settingsCodeFontFamily => 'コードフォント';
 
   @override
+  String get settingsEditorFontFamily => '本文フォント';
+
+  @override
   String get settingsEditorMaxWidth => 'エディタの最大幅';
 
   @override