+++ code/lib/core/i18n/l10n/app_localizations_pt.dart
@@ -402,6 +402,9 @@ class AppLocalizationsPt extends AppLocalizations {
   String get settingsCodeFontFamily => 'Fonte de código';
 
   @override
+  String get settingsEditorFontFamily => 'Fonte do editor';
+
+  @override
   String get settingsEditorMaxWidth => 'Largura máxima do editor';
 
   @override